Type sets the strol to output an int

+++ 0x0A-argc_argv/4-add.c
@@ -22,7 +22,7 @@ int main(int argc, char **argv)
 
 	for (a = 1; a < argc; a++)
 	{
-		num = strtol(argv[a], &endptr, 10);
+		num = (int)strtol(argv[a], &endptr, 10);
 
 	if (*endptr != '\0')
 	{
